test: rename e2e spec for notification flow
