+++ utils/date.ts
@@ -1,0 +1,3 @@
+export function formatDateToDutchLocale(dotDate: string): string {
+  return dotDate.replace(/\./g, '-');
+}
